resolve async cleanup
- remove `struct Curl_dns_entry *dns` entry
- remove `done` BIT
- make port an uint16_t
- make ip_version an uint8_t
- rename `Curl_async_is_resolved()` to `Curl_async_take_result()`,
  returning CURLE_AGAIN while waiting on the result
- remove `Curl_once_resolved()`

+++ lib/curl_addrinfo.c
@@ -404,7 +404,7 @@ static CURLcode ip2addr(struct Curl_addrinfo **addrp, int af,
  * Given an IPv4 or IPv6 dotted string address, this converts it to a proper
  * allocated Curl_addrinfo struct and returns it.
  */
-CURLcode Curl_str2addr(const char *dotted, int port,
+CURLcode Curl_str2addr(const char *dotted, uint16_t port,
                        struct Curl_addrinfo **addrp)
 {
   struct in_addr in;
